requirements.txt
langgraph
langchain
langchain-google-genai
google-cloud-vision
opencv-python
python-dotenv
Pillow
opencv-python
streamlit
streamlit-mic-recorder
sounddevice
scipy
noisereduce
openai-whisper
edge-tts
pydub
webrtcvad-wheels
faster_whisper